Consolidate !table commands into !compare.
I've settled on a preformatted text table as the form that
works best on different platforms. I've also decided on
having a single command for simplicity to avoid confusing
users.

My testing seems to show that on mobile devices where
the table is too wide it will show fine once the user rotates
to landscape view.

I think !compare is perhaps a clearer description of what
the command does for non-technical users, so that's the
reason for the rename there.